Phase 2 step 4: Extract orders_io + stats
